+++ src/Entity/Doctrine/ResultConfiguration.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Paysera\Pagination\Entity\Doctrine;
+
+class ResultConfiguration extends ConfiguredQuery
+{
+}